+++ src/interactive_menu.cpp
@@ -43,8 +43,8 @@ void interactive_menu::resize(uint16_t width, uint16_t height)
   wresize(menu_window, height, width);
   this->width = width;
   this->height = height;
-  uint16_t body_size = height-header.size();
-  uint16_t max_increment = lines.size() > body_size ? lines.size()-body_size : 0;
+  uint16_t body_size = height-static_cast<uint16_t>(header.size());
+  uint16_t max_increment = static_cast<uint16_t>(lines.size()) > body_size ? static_cast<uint16_t>(lines.size())-body_size : 0;
   if(position > max_increment)
   {
     position = max_increment;
@@ -64,19 +64,19 @@ void interactive_menu::reset_position()
 void interactive_menu::print()
 {
   werase(menu_window);
-  for(size_t i = 0; i < header.size(); ++i)
+  for(uint16_t i = 0; i < header.size(); ++i)
   {
     mvwaddstr(menu_window, i, 0, header[i].c_str());
   }
-  for(size_t i = 0; i + position < lines.size() && i + header.size() < height; ++i)
+  for(uint16_t i = 0; i + position < lines.size() && i + header.size() < height; ++i)
   {
-    if(header.size() + i == static_cast<size_t>(height)-1 && i + position + 1 < lines.size())
+    if(static_cast<uint16_t>(header.size()) + i == height-1 && i + position + 1u < lines.size())
     {
-      mvwaddstr(menu_window, i+header.size(), 0, "** MORE **");
+      mvwaddstr(menu_window, i+static_cast<uint16_t>(header.size()), 0, "** MORE **");
     }
     else
     {
-      mvwaddstr(menu_window, i+header.size(), 0, lines[i + position].c_str());
+      mvwaddstr(menu_window, i+static_cast<uint16_t>(header.size()), 0, lines[i + position].c_str());
     }
   }
   wnoutrefresh(menu_window);
@@ -95,14 +95,14 @@ int interactive_menu::get_player_input()
     }
     else if(player_input == KEY_DOWN)
     {
-      uint16_t body_size = height-header.size();
-      uint16_t max_increment = lines.size() > body_size ? lines.size()-body_size : 0;
+      uint16_t body_size = static_cast<uint16_t>(height-header.size());
+      uint16_t max_increment = lines.size() > body_size ? static_cast<uint16_t>(lines.size())-body_size : 0;
       position = position + 1 >= max_increment ? max_increment : position + 1;
     }
     else if(player_input == ' ')
     {
-      uint16_t body_size = height-header.size();
-      uint16_t max_increment = lines.size() > body_size ? lines.size()-body_size : 0;
+      uint16_t body_size = height-static_cast<uint16_t>(header.size());
+      uint16_t max_increment = lines.size() > body_size ? static_cast<uint16_t>(lines.size())-body_size : 0;
       position = position + body_size >= max_increment ? max_increment : position + body_size;
     }
     else
